% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Title fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Number of a state}

\item{year}{The year in question}
}
\value{
A plot or set of plots based
 on latitude and longitude from the data file
}
\description{
Receive a state number and one year.
make the appropriate file name using the year
and the make_filename function obtains a data frame from fars_read()
}
\details{
Error checks to make sure the state number exists

Uses make_filename(year)
     fars_read(filename)
}
\examples{
# We set the working directory where the compressed files are stored within
# the package.
x <- setwd(system.file("extdata", package = "packagefars"))
setwd(x)
}
