% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Title fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{of one or more years}
}
\value{
Creates one or more datasets based on year number.
Returns NULL if there is an error
}
\description{
This function receives a list of years (one or several years)
and calls the make_filename()
function for each year and then fills those files with
the corresponding year data from the main data set.
}
\details{
if one year is not valid an error will
be generated and the function will stop

Uses make_filename(year)
}
\examples{
\dontrun{x <- setwd(system.file("extdata", package = "packagefars"))}
# Data with the car fatalities in 2013.
\dontrun{fars_read_years(2013)}
# Data with the car fatalities in 2013 and 2014.
\dontrun{fars_read_years(c(2013, 2014))}
\dontrun{setwd(x)}
}
