% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Title fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{One or more years, no error checking}
}
\value{
A data frame of counts by month and year
}
\description{
Receive a list of years (one or several years) and
pass it to the function fars_read_years ().
This function receives a set of data with the month and year columns.
It uses several functions of dplyr to count the number of observations
per month for each year.
}
\details{
Uses fars_read_years(years)
}
\examples{

\dontrun{x <- setwd(system.file("extdata", package = "packagefars"))}
# Summary with the tallies of car fatalities in 2013.
\dontrun{fars_summarize_years(2013)}
# Summary with the tallies of car fatalities in 2013 and 2014.
\dontrun{fars_summarize_years(c(2013, 2014))}
\dontrun{setwd(x)}
}
