% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Title make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{as a string}
}
\value{
a filename in the format 'accident_year.csv.bz2'. The function prints the file name.
}
\description{
This function allows you to create accident data files per year
receive the file in csv.bz2 format
}
\examples{
\dontrun{x <- setwd(system.file("extdata", package = "packagefars"))}
\dontrun{make_filename(2013)}
\dontrun{setwd(x)}
}
